requirements.txt
pydantic~=2.6.4
fastapi~=0.110.0
bcrypt~=4.1.2
PyJWT~=2.8.0
passlib~=1.7.4